refactor: move the rebase instruction to the end
Why? In the lecture of Wednesday, 25th of November 2020, I couldn't
explain how to do git operations like `git reset`, `git cherry-pick`,
`git rebase` or `git rebase --interactive`.

So let's move this to the end of the homework instructions. Next
lecture on December 2nd we will cover that.

It's fine to do the rebase as a last step.